+++ CHANGELOG.md
@@ -16,6 +16,11 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+  
+### :bug: Bug Fix
+
+- `ricos-content`
+  - [#3019](https://github.com/wix/ricos/pull/3019) fromDraft: Invalid entity key protected
 
 ### :nail_care: Polish
 
@@ -24,6 +29,7 @@
 - `vertical-embed`
   - [#3005](https://github.com/wix/ricos/pull/3005) vertical-embed modal new design
 - `line-spacing`
+
   - [#3009](https://github.com/wix/ricos/pull/3009) add custom line-height selection to the line-heights options display
 - `ricos-content`
   - [#3016](https://github.com/wix/ricos/pull/3016) Polls Content Builder API support
@@ -32,6 +38,8 @@
 
 - `schema`
   - [#3013](https://github.com/wix/ricos/pull/3013) undefined fields added to plugin defaults JSON
+- `divider`
+  - [#3017](https://github.com/wix/ricos/pull/3017) set divider width to 100% and remove editor bounds consideration
 
 </details>
 <hr/>